+++ internal/internal_workflow_testsuite.go
@@ -52,7 +52,6 @@ const (
 	defaultTestWorkflowTypeName = "default-test-workflow-type-name"
 	defaultTestDomainName       = "default-test-domain-name"
 	workflowTypeNotSpecified    = "workflow-type-not-specified"
-	testHeader                  = "test-header"
 )
 
 type (
@@ -284,7 +283,6 @@ func newTestWorkflowEnvironmentImpl(s *WorkflowTestSuite) *testWorkflowEnvironme
 	if env.workerOptions.DataConverter == nil {
 		env.workerOptions.DataConverter = getDefaultDataConverter()
 	}
-	env.workerOptions.ContextPropagators = append(env.workerOptions.ContextPropagators, &TestContextPropagator{keys: []string{testHeader}})
 
 	return env
 }
@@ -406,16 +404,12 @@ func (env *testWorkflowEnvironmentImpl) executeWorkflowInternal(delayStart time.
 		panic(err)
 	}
 	env.workflowDef = workflowDefinition
-	header := &shared.Header{
-		Fields: map[string][]byte{
-			"test-header": []byte("test-data"),
-		},
-	}
+
 	// env.workflowDef.Execute() method will execute dispatcher. We want the dispatcher to only run in main loop.
 	// In case of child workflow, this executeWorkflowInternal() is run in separate goroutinue, so use postCallback
 	// to make sure workflowDef.Execute() is run in main loop.
 	env.postCallback(func() {
-		env.workflowDef.Execute(env, header, input)
+		env.workflowDef.Execute(env, nil, input)
 		// kick off first decision task to start the workflow
 		if delayStart == 0 {
 			env.startDecisionTask()
@@ -470,11 +464,6 @@ func (env *testWorkflowEnvironmentImpl) executeActivity(
 		},
 		ActivityType: *activityType,
 		Input:        input,
-		Header: &shared.Header{
-			Fields: map[string][]byte{
-				testHeader: []byte("test-data"),
-			},
-		},
 	}
 
 	task := newTestActivityTask(
