+++ workflow/context_propagator.go
@@ -20,12 +20,7 @@
 
 package workflow
 
-import (
-	"context"
-	"fmt"
-
-	"go.uber.org/cadence/internal"
-)
+import "go.uber.org/cadence/internal"
 
 type (
 	// HeaderReader is an interface to read information from cadence headers
@@ -37,74 +32,9 @@ type (
 	// ContextPropagator is an interface that determines what information from
 	// context to pass along
 	ContextPropagator = internal.ContextPropagator
+
+	// StringMapPropagator is a ContextPropagator that propagates a list of keys
+	// across a workflow, interpreting the payloads as strings. This is provided
+	// as a sample for other context propagator implementations.
+	StringMapPropagator = internal.StringMapPropagator
 )
-
-// ContextKey to be used for the StringMapPropagator
-type ContextKey string
-
-// stringMapPropagator propagates the list of keys across a workflow,
-// interpreting the payloads as strings. This is provided as a sample for other
-// context propagator implementations.
-type stringMapPropagator struct {
-	keys map[string]struct{}
-}
-
-// NewStringMapPropagator returns a context propagator that propagates a set of
-// string key-value pairs across a workflow
-func NewStringMapPropagator(keys []string) ContextPropagator {
-	keyMap := make(map[string]struct{}, len(keys))
-	for _, key := range keys {
-		keyMap[key] = struct{}{}
-	}
-	return &stringMapPropagator{keyMap}
-}
-
-// Inject injects values from context into headers for propagation
-func (s *stringMapPropagator) Inject(ctx context.Context, writer HeaderWriter) error {
-	for key := range s.keys {
-		value, ok := ctx.Value(ContextKey(key)).(string)
-		if !ok {
-			return fmt.Errorf("unable to extract key from context %v", key)
-		}
-		writer.Set(key, []byte(value))
-	}
-	return nil
-}
-
-// InjectFromWorkflow injects values from context into headers for propagation
-func (s *stringMapPropagator) InjectFromWorkflow(ctx Context, writer HeaderWriter) error {
-	for key := range s.keys {
-		value, ok := ctx.Value(ContextKey(key)).(string)
-		if !ok {
-			return fmt.Errorf("unable to extract key from context %v", key)
-		}
-		writer.Set(key, []byte(value))
-	}
-	return nil
-}
-
-// Extract extracts values from headers and puts them into context
-func (s *stringMapPropagator) Extract(ctx context.Context, reader HeaderReader) (context.Context, error) {
-	if err := reader.ForEachKey(func(key string, value []byte) error {
-		if _, ok := s.keys[key]; ok {
-			ctx = context.WithValue(ctx, ContextKey(key), string(value))
-		}
-		return nil
-	}); err != nil {
-		return nil, err
-	}
-	return ctx, nil
-}
-
-// ExtractToWorkflow extracts values from headers and puts them into context
-func (s *stringMapPropagator) ExtractToWorkflow(ctx Context, reader HeaderReader) (Context, error) {
-	if err := reader.ForEachKey(func(key string, value []byte) error {
-		if _, ok := s.keys[key]; ok {
-			ctx = WithValue(ctx, ContextKey(key), string(value))
-		}
-		return nil
-	}); err != nil {
-		return nil, err
-	}
-	return ctx, nil
-}
